\name{dataqol2}
\alias{dataqol2}
\docType{data}
\title{
Longitudinal quality of life data
}
\description{
 A data frame with 6 quality of life measures for 60 patients. The dataset is in long format.
}
\usage{data(dataqol2)}
\format{

  \describe{
    \item{\code{id}}{subject identification number}
    \item{\code{time}}{visit number for quality of life assessment }
    \item{\code{date}}{date of quality of life measurement}
    \item{\code{QoL}}{score of global quality of life}
    \item{\code{pain}}{score of pain}
    \item{\code{arm}}{treatment arm}
    \item{\code{death}}{date of death}
  }
}


