\name{whicha}
\alias{whicha}
\title{
To check a condition
}
\description{
A function to check a condition
}
\usage{
whicha(condition)
}
\arguments{
  \item{condition}{
a logical vector
}
}

\value{
0 if the logical vector does not contain the value \code{TRUE}. Otherwise give the \code{TRUE} indices
}
